+++ CHANGELOG.md
@@ -20,7 +20,12 @@ Given a version number `MAJOR.MINOR.PATCH`, we increment the:
 - Backward compatibility in versions `0.0.z` is **not guaranteed** when `z` is increased.
 - Backward compatibility in versions `0.y.z` is **not guaranteed** when `y` is increased.
 
-# Unreleased
+## Unreleased
+
+### Added
+
+- Add support for Single-Sign-On login.
+  - Use `terramate cloud login --sso` for SSO enabled organizations.
 
 ## v0.12.1